Split argument parsing into a module.
For integrating with the atomic command.  atomic will try to import
the module and call add_subparsers().  The module handles the rest.

N.B. This makes NO changes to classes moved from client_script.py.
